[Bard] various changes and updates (#201)
## Description

Closes #26 #29 

## Checklist
- [ ] Updated CHANGELOG files
- [ ] Updated Documentation
- [ ] Unit Tests Created
- [ ] php-cs-fixer

+++ src/SonsOfPHP/Bard/src/Console/Command/AddCommand.php
@@ -24,26 +24,49 @@ final class AddCommand extends AbstractCommand
         $this
             ->setDescription('Add new repo')
             ->addOption('branch', null, InputOption::VALUE_REQUIRED, 'What branch we working with?', 'main')
+            ->addOption('dry-run', null, InputOption::VALUE_NONE, 'Dry Run (Do not make any changes)')
             ->addArgument('path', InputArgument::REQUIRED, 'Path where code will be')
-            ->addArgument('repository', InputArgument::REQUIRED, 'Repository')
+            ->addArgument('repository', InputArgument::REQUIRED, 'Repository Uri')
+            ->setHelp(
+                <<<'HELP'
+The <info>add</info> command will add additional repositories that need to be managed
+into the `bard.json` file.
+
+Examples:
+
+    <comment>%command.full_name% src/SonsOfPHP/Bard git@github.com:vendor/package.git</comment>
+
+Read more at https://docs.sonsofphp.com/bard/
+HELP
+            )
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $bardConfig = new JsonFile($input->getOption('working-dir') . '/bard.json');
-        $formatter  = $this->getHelper('formatter');
         $io         = new SymfonyStyle($input, $output);
+        $isDryRun   = $input->getOption('dry-run');
 
+        // ---
         if (null === $packages = $bardConfig->getSection('packages')) {
             $packages = [];
         }
-
-        // @todo Check to make sure this package does not already exist
+        foreach ($packages as $pkg) {
+            if ($pkg['path'] === $input->getArgument('path')) {
+                $io->error([
+                    sprintf('It appears that the path "%s" is currently being used.', $pkg['path']),
+                    'Please check your bard.json file',
+                ]);
+                return self::FAILURE;
+            }
+        }
         $packages[] = [
-            'path' => $input->getArgument('path'),
+            'path'       => $input->getArgument('path'),
             'repository' => $input->getArgument('repository'),
         ];
+        // ---
+
         $bardConfig = $bardConfig->setSection('packages', $packages);
         $commands = [
             ['git', 'subtree', 'add', '--prefix', $input->getArgument('path'), $input->getArgument('repository'), $input->getOption('branch'), '--squash'],
@@ -51,10 +74,15 @@ final class AddCommand extends AbstractCommand
         foreach ($commands as $cmd) {
             $process = new Process($cmd);
             $io->text($process->getCommandLine());
-            $this->getHelper('process')->mustRun($output, $process, sprintf('There was and error running command: %s', $process->getCommandLine()));
+            if (!$isDryRun) {
+                $this->getHelper('process')->mustRun($output, $process, sprintf('There was and error running command: %s', $process->getCommandLine()));
+            }
         }
 
-        file_put_contents($bardConfig->getFilename(), $bardConfig->toJson());
+        if (!$isDryRun) {
+            file_put_contents($bardConfig->getFilename(), $bardConfig->toJson());
+        }
+
         $io->success('Package has been added.');
 
         return self::SUCCESS;
